Refactor Slack notification to analyzer helper file
